refactor(hq): extract order detail view cards into components (pure moves)
Move the view-mode JSX of the order detail page into
src/components/ops/orders/detail/ (DetailHeader, StatusStepperCard,
DeliveryCard, ItemsCard, PaymentCard + shared helpers/types). The page
keeps all state, handlers, edit mode, dialogs, and the Print View
untouched — zero visual or behavioral change.
